LPD-78636 Move home panel apps to correct package
